+++ unit.test.js
@@ -5,23 +5,22 @@ const cloudevent = new Cloudevent({
 	data: JSON.stringify({ some: 'payload' }),
 	source: 'https://github.com/1mill/cloudevents',
 	type: 'cmd.do-this-command.v0',
-	originatorid: 'user.id.1234',
-})
+}).originator({ originatorid: 'user.id.1234' })
 
 console.log(cloudevent)
 // {
-//   id: 'tEzBztYRlPAaGY3uWIVWI',
+//   id: 'g-wea_sYvxmvsGeL3YCSM',
 //   source: 'https://github.com/1mill/cloudevents',
 //   type: 'cmd.do-this-command.v0',
 //   specversion: '1.0',
-//   time: '2022-09-21T03:58:36.995Z',
+//   time: '2022-09-21T04:27:10.729Z',
 //   data: '{"some":"payload"}',
 //   datacontenttype: 'application/json',
 //   dataschema: undefined,
 //   subject: undefined,
-//   originid: 'tEzBztYRlPAaGY3uWIVWI',
+//   originid: 'g-wea_sYvxmvsGeL3YCSM',
 //   originsource: 'https://github.com/1mill/cloudevents',
-//   origintime: '2022-09-21T03:58:36.995Z',
+//   origintime: '2022-09-21T04:27:10.729Z',
 //   origintype: 'cmd.do-this-command.v0',
 //   originatorid: 'user.id.1234',
 //   wschannelid: undefined
@@ -33,23 +32,24 @@ const enrichedCloudevent = new Cloudevent({
 	type: 'fct.this-thing-happened.v0',
 })
 .origin({ cloudevent })
+.originator({ cloudevent })
 .wschannel({ wschannelid: 'some-prefix:my-resource-name#id=12345' })
 
 console.log(enrichedCloudevent)
 // {
-//   id: '4_7YyYMjm-YPE3f20B1Ow',
+//   id: 'zojPIW1eMm_JsvQI5VnfR',
 //   source: 'https://www.erikekberg.com/',
 //   type: 'fct.this-thing-happened.v0',
 //   specversion: '1.0',
-//   time: '2022-09-21T03:58:37.005Z',
+//   time: '2022-09-21T04:27:10.740Z',
 //   data: '{"new":"payload","value":true}',
 //   datacontenttype: 'application/json',
 //   dataschema: undefined,
 //   subject: undefined,
-//   originid: 'tEzBztYRlPAaGY3uWIVWI',
+//   originid: 'g-wea_sYvxmvsGeL3YCSM',
 //   originsource: 'https://github.com/1mill/cloudevents',
-//   origintime: '2022-09-21T03:58:36.995Z',
+//   origintime: '2022-09-21T04:27:10.729Z',
 //   origintype: 'cmd.do-this-command.v0',
-//   originatorid: undefined,
+//   originatorid: 'user.id.1234',
 //   wschannelid: 'some-prefix:my-resource-name#id=12345'
 // }